+++ CHANGELOG.rst
@@ -6,8 +6,15 @@ Version `dev`_
 ================
 **Date**: unreleased
 
+* API client:
+
+  * Add "include_invalid" option to QUICK lookup to return invalid IPs as part of the JSON response
+
+  * Added support for new /riot endpoint
+
 * CLI:
 
+  * Added support for new riot command
   * Updated json_formatter for query commands to return data only as New Line Delimited JSON
 
 * API client: